Fix bugs for ldc long
Former-commit-id: d7ffa4ecdcb65429fa3dce9f2bbe67cbef1effe5

+++ MIB/src/edu/columbia/psl/cc/inst/DynamicMethodMiner.java
@@ -528,7 +528,7 @@ public class DynamicMethodMiner extends MethodVisitor {
 	@Override
 	public void visitLdcInsn(Object cst) {
 		if (this.shouldInstrument() && !this.constructor) {
-			if (cst instanceof Double || cst instanceof Float) {
+			if (cst instanceof Double || cst instanceof Long) {
 				this.handleLdc(Opcodes.LDC, 2, cst.toString());
 			} else {
 				this.handleLdc(Opcodes.LDC, 1, cst.toString());
